Move localization resources to lib/intl.
