+++ lib/extensions/country_extensions.dart
@@ -1,4 +1,4 @@
-import 'package:bratacha/country_localizations.dart';
+import 'package:bratacha/intl/country_localizations.dart';
 import 'package:bratacha/modules/country_database/country_database.dart';
 
 extension CountryExtensions on Country {
